Move DocPropertyWriter to document package.
It requires a document only and thus should be in the opengever.document
package.
