+++ opengever/document/tests/test_versioner.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from ftw.testbrowser import browsing
+from opengever.document.docprops import DocPropertyWriter
 from opengever.document.versioner import Versioner
-from opengever.dossier.docprops import DocPropertyWriter
 from opengever.testing import IntegrationTestCase
 from plone.namedfile.file import NamedBlobFile
 from zope.filerepresentation.interfaces import IRawWriteFile
